+++ src/URDFLoader.d.ts
@@ -9,6 +9,26 @@ interface MeshLoadFunc{
     (url: string, manager: LoadingManager, onLoad: MeshLoadDoneFunc): void;
 }
 
+interface RobotHookFunc{
+    (robotNodes: Element[], linkMap: { [key: string]: Object3D }, jointMap: { [key: string]: Object3D }, curRobotObject3D: Object3D): void;
+}
+
+interface JointHookFunc{
+    (curJointChildNode: Element, curJointObject3D: Object3D): void;
+}
+
+interface LinkHookFunc{
+    (curLinkChildNode: Element, curLinkObject3D: Object3D): void;
+}
+
+interface LinkVisualHookFunc{
+    (curLinkVisualNode: Element, curLinkVisualObject3D: Object3D): void;
+}
+
+interface LinkCollisionHookFunc{
+    (curLinkCollisionNode: Element, curLinkCollisionObject3D: Object3D): void;
+}
+
 export default class URDFLoader {
 
     manager: LoadingManager;
@@ -21,6 +41,11 @@ export default class URDFLoader {
     parseCollision: boolean;
     packages: string | { [key: string]: string } | ((targetPkg: string) => string);
     loadMeshCb: MeshLoadFunc;
+    robotHook: RobotHookFunc;
+    jointHook: JointHookFunc;
+    linkHook: LinkHookFunc;
+    linkVisualHook: LinkVisualHookFunc;
+    linkCollisionHook: LinkCollisionHookFunc;
 
     constructor(manager?: LoadingManager);
     loadAsync(urdf: string): Promise<URDFRobot>;